Rename MapEvent file
